add error handling

+++ src/services/products-service.js
@@ -1,4 +1,5 @@
 import { prisma } from "../database.js"
+import ResponseError from "../utils/response-error.js"
 
 const create = async (product) => {
   return await prisma.product.create({
@@ -14,18 +15,34 @@ const create = async (product) => {
 }
 
 const getAll = async () => {
-  return await prisma.product.findMany()
+  const result = await prisma.product.findMany()
+  if (result.length === 0) {
+    throw new ResponseError(200, 'There is no products')
+  }
+  return result
 }
 
 const getById = async (productId) => {
-  return await prisma.product.findMany({
+  const result = await prisma.product.findMany({
     where: {
       id: productId
     }
   })
+  if (result.length === 0) {
+    throw new ResponseError(404, 'Product is not found')
+  }
+  return result
 }
 
 const update = async (product) => {
+  const result = await prisma.product.count({
+    where: {
+      id: product.id
+    }
+  })
+  if (result !== 1) {
+    throw new ResponseError(404, 'Product is not found')
+  }
   return await prisma.product.update({
     data: product,
     where: {
@@ -42,6 +59,14 @@ const update = async (product) => {
 }
 
 const remove = async (productId) => {
+  const result = await prisma.product.count({
+    where: {
+      id: product.id
+    }
+  })
+  if (result !== 1) {
+    throw new ResponseError(404, 'Product is not found')
+  }
   return await prisma.product.delete({
     where: {
       id: productId
